requirements.txt
telethon
python-dotenv
sqlalchemy
psycopg2-binary
